Create peopleChoices
